techtest/requirements.txt
django
django-extensions
django-filter
django-tagulous
djangorestframework
ipython
markdown